electionnews/requirements.txt
feedparser
pandas
PyYAML
requests
transformers
torch***
